requirements.txt
annotated-types==0.7.0
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
numpy==2.3.2
pandas==2.3.1
pdfminer.six==20250506
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
psycopg[binary]>=3.1
google-api-python-client>=2.137.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
imapclient>=3.0
